Add reward discount functionality to bookings and update related models

+++ database/migrations/2025_07_09_121131_create_bookings_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
                 ->onDelete('cascade');
             $table->timestamp('booked_at')->useCurrent();
             $table->enum('status', ['pending', 'confirmed', 'canceled'])->default('pending');
-            $table->integer('total_price')->default(0);
+            $table->decimal('total_price', 8, 2)->default(0.00);
             $table->integer('total_points')->default(0);
             $table->integer('ticket_quantity')->default(1);
             $table->softDeletes();
